move DreamsTagChart into charts
